+++ app/src/main/java/com/aiwazian/messenger/data/SearchInfo.kt
@@ -4,7 +4,7 @@ import androidx.annotation.Keep
 
 @Keep
 data class SearchInfo(
-    @Keep val chatId: Int,
+    @Keep val chatId: Long,
     @Keep val name: String,
     @Keep val publicLink: String
 )
