+++ app/src/main/java/com/aiwazian/messenger/data/UserInfo.kt
@@ -5,7 +5,7 @@ import com.aiwazian.messenger.interfaces.Profile
 
 @Keep
 data class UserInfo(
-    @Keep override var id: Int = 0,
+    @Keep override var id: Long = 0,
     @Keep var firstName: String = "",
     @Keep var lastName: String = "",
     @Keep var username: String? = null,
