+++ app/src/main/java/com/aiwazian/messenger/database/entity/AccountEntity.kt
@@ -4,4 +4,4 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity("account")
-data class AccountEntity(@PrimaryKey val id: Int, val isCurrent: Boolean)
+data class AccountEntity(@PrimaryKey val id: Long, val isCurrent: Boolean)
