+++ app/src/main/java/com/aiwazian/messenger/database/entity/ChannelEntity.kt
@@ -5,10 +5,10 @@ import androidx.room.PrimaryKey
 
 @Entity("channel")
 data class ChannelEntity(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     var name: String,
     var bio: String = "",
-    val ownerId: Int,
+    val ownerId: Long,
     val subscribers: Int,
     val removedUser: Int,
     val channelType: Int,
