+++ app/src/main/java/com/aiwazian/messenger/database/entity/MessageEntity.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 @Entity("message")
 data class MessageEntity(
     @PrimaryKey var id: Int,
-    val senderId: Int,
-    val chatId: Int,
+    val senderId: Long,
+    val chatId: Long,
     val text: String = "",
     val sendTime: Long = 0,
     var isRead: Boolean = false
